+++ modules/Student_Billing/includes/DailyTotals.php
@@ -23,7 +23,7 @@ DrawHeader( _( 'Report Timeframe' ) . ': ' .
 echo '</form>';
 
 $billing_payments = DBGetOne( "SELECT sum(AMOUNT) AS AMOUNT
-	FROM BILLING_PAYMENTS
+	FROM billing_payments
 	WHERE SYEAR='" . UserSyear() . "'
 	AND SCHOOL_ID='" . UserSchool() . "'
 	AND PAYMENT_DATE BETWEEN '" . $start_date . "'